Major changes
Most of this stuff isn't even finished half way yet. Some files might be completely unneccessary or won't be needed in the future. However, I wanted to make sure to have a commit after soo many iterations. There's still a lot left to do, but I'm doing this project for fun and currently fine with just working on it every few weeks :)